+++ pkg/scalers/kafka_scaler.go
@@ -395,7 +395,7 @@ func (s *kafkaScaler) GetMetrics(ctx context.Context, metricName string, metricS
 
 	metric := external_metrics.ExternalMetricValue{
 		MetricName: metricName,
-		Value:      *resource.NewQuantity(int64(totalLag), resource.DecimalSI),
+		Value:      *resource.NewQuantity(totalLag, resource.DecimalSI),
 		Timestamp:  metav1.Now(),
 	}
